Moving candidate files to folder and adding dialog
