+++ frontend/src/components/ElectionForm/Candidates/PhotoCropper.js
